1. Move `getIconFromManifestVariable` in WoosmapProvider
