+++ woosmapgeofencingcore/src/main/java/com/webgeoservices/woosmapgeofencingcore/WoosmapCore.java
@@ -267,19 +267,4 @@ public class WoosmapCore extends WoosmapProvider{
         });
     }
 
-    protected int getIconFromManifestVariable() {
-        ApplicationInfo mApplicationInfo;
-        try {
-            mApplicationInfo = context.getPackageManager().getApplicationInfo(context.getPackageName(), PackageManager.GET_META_DATA);
-            Bundle bundle = mApplicationInfo.metaData;
-            if (bundle.containsKey("woosmap.messaging.default_notification_icon")) {
-                return bundle.getInt("woosmap.messaging.default_notification_icon", R.drawable.ic_local_grocery_store_black_24dp);
-            } else {
-                return R.drawable.ic_local_grocery_store_black_24dp;
-            }
-        } catch (PackageManager.NameNotFoundException e) {
-            e.printStackTrace();
-            return R.drawable.ic_local_grocery_store_black_24dp;
-        }
-    }
 }
